+++ openedx/core/djangoapps/coursegraph/apps.py
@@ -12,3 +12,5 @@ class CoursegraphConfig(AppConfig):
     AppConfig for courseware app
     """
     name = 'openedx.core.djangoapps.coursegraph'
+
+    from . import tasks  # pylint: disable=unused-variable
